+++ src/main/java/com/smoorbuilderserver/repository/CombatActionDescriptionRepository.java
@@ -1,6 +1,5 @@
 package com.smoorbuilderserver.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -9,13 +8,13 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.CombatActionDescription;
 
 @Repository
-public interface CombatActionDescriptionRepository extends CrudRepository<CombatActionDescription, BigInteger> {
+public interface CombatActionDescriptionRepository extends CrudRepository<CombatActionDescription, Long> {
 	
 	public CombatActionDescription findByActionName(String actionName);
 	
 	public CombatActionDescription findByActionDescription(String actionDescription);
 	
-	public CombatActionDescription findByid(BigInteger id);
+	public CombatActionDescription findByid(Long id);
 	
 	public List<CombatActionDescription> findDistinctCombatActionDescriptionByActionName(String actionName);
 }
