+++ src/main/java/com/smoorbuilderserver/repository/GeneralSkillDescriptionRepository.java
@@ -1,6 +1,5 @@
 package com.smoorbuilderserver.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -9,11 +8,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.GeneralSkillDescription;
 
 @Repository
-public interface GeneralSkillDescriptionRepository extends CrudRepository<GeneralSkillDescription, BigInteger> {
+public interface GeneralSkillDescriptionRepository extends CrudRepository<GeneralSkillDescription, Long> {
 	
 	public GeneralSkillDescription findByGeneralSkillName(String generalSkillName);
 	
-	public GeneralSkillDescription findByid(BigInteger id);
+	public GeneralSkillDescription findByid(Long id);
 	
 	public List<GeneralSkillDescription> findDistinctGeneralSkillDescriptionByGeneralSkillName(String generalSkillName);
 }
