+++ src/main/java/com/smoorbuilderserver/repository/SpellcastingDescriptionRepository.java
@@ -1,6 +1,5 @@
 package com.smoorbuilderserver.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -9,11 +8,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.SpellcastingDescription;
 
 @Repository
-public interface SpellcastingDescriptionRepository extends CrudRepository<SpellcastingDescription, BigInteger> {
+public interface SpellcastingDescriptionRepository extends CrudRepository<SpellcastingDescription, Long> {
 	
 	public SpellcastingDescription findBySpellcastingName(String generalSkillName);
 	
-	public SpellcastingDescription findByid(BigInteger id);
+	public SpellcastingDescription findByid(Long id);
 	
 	public List<SpellcastingDescription> findDistinctSpellcastingDescriptionBySpellcastingName(String spellcastingDescriptionName);
 }
