+++ src/main/java/com/smoorbuilderserver/repository/UserRepository.java
@@ -1,6 +1,5 @@
 package com.smoorbuilderserver.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -9,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.User;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, BigInteger> {
+public interface UserRepository extends CrudRepository<User, Long> {
 	
 	public User findByUsername(String username);
